requirements.txt
pytest==7.4.0
requests==2.31.0
pytest-html==4.1.1
allure-pytest==2.13.2
python-dotenv==1.0.0